refactor(api): use shared not found error
