+++ api/src/evaluation/domain/usecases/get-answer.js
@@ -1,4 +1,4 @@
-import { NotFoundError } from '../../../../lib/domain/errors.js';
+import { NotFoundError } from '../../../shared/domain/errors.js';
 
 const getAnswer = async function ({ answerId, userId, answerRepository, assessmentRepository } = {}) {
   const integerAnswerId = parseInt(answerId);
