+++ api/tests/evaluation/integration/infrastructure/repositories/autonomous-course-target-profile-repository_test.js
@@ -1,7 +1,7 @@
 import { constants } from '../../../../../lib/domain/constants.js';
-import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { AutonomousCourseTargetProfile } from '../../../../../src/evaluation/domain/models/AutonomousCourseTargetProfile.js';
 import { repositories } from '../../../../../src/evaluation/infrastructure/repositories/index.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Integration | Repository | Autonomous Course Target Profile', function () {
