+++ api/tests/evaluation/unit/domain/usecases/get-answer_test.js
@@ -1,5 +1,5 @@
-import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { getAnswer } from '../../../../../src/evaluation/domain/usecases/get-answer.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | get-answer', function () {
